şişman initrd desteği

+++ src/addons/00-modules.sh
@@ -7,13 +7,17 @@
 debug "Kernel Version" "${KERNELVER}"
 debug "Module Directory" "${MODDIR}"
 mkdir -p ${WORKDIR}/${MODDIR}
-debug "Install main modules"
-cp -prf ${MODDIR}/kernel/{crypto,fs,lib} ${WORKDIR}/${MODDIR}
-debug "Install drivers"
-cp -prf ${MODDIR}/kernel/drivers/{block,ata,md,firewire} ${WORKDIR}/${MODDIR}
-cp -prf ${MODDIR}/kernel/drivers/{scsi,pcmcia,virtio} ${WORKDIR}/${MODDIR}
-cp -prf ${MODDIR}/kernel/drivers/usb/{host,storage} ${WORKDIR}/${MODDIR}
-if [ "$debug" != "true" ] ; then
-	debug "Remove debug drivers"
-	find ${WORKDIR}/${MODDIR} | grep debug | xargs rm -rf
+if [ "$allmodule" == true ] && [ "$skipglibc" != "true" ] && [ "$skipudev" != "true" ]; then
+	cp -prf ${MODDIR}/* ${WORKDIR}/${MODDIR}
+else
+	debug "Install main modules"
+	cp -prf ${MODDIR}/kernel/{crypto,fs,lib} ${WORKDIR}/${MODDIR}
+	debug "Install drivers"
+	cp -prf ${MODDIR}/kernel/drivers/{block,ata,md,firewire} ${WORKDIR}/${MODDIR}
+	cp -prf ${MODDIR}/kernel/drivers/{scsi,pcmcia,virtio} ${WORKDIR}/${MODDIR}
+	cp -prf ${MODDIR}/kernel/drivers/usb/{host,storage} ${WORKDIR}/${MODDIR}
+	if [ "$debug" != "true" ] ; then
+		debug "Remove debug drivers"
+		find ${WORKDIR}/${MODDIR} | grep debug | xargs rm -rf
+	fi
 fi
